app/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic-settings
pydantic[email]
asyncpg
supabase
pyrebase4
firebase-admin
requests
python-jose[cryptography]
